[FEATURE] 응답 데이터 구조 및 버그 수정

+++ src/main/java/com/knu/sosuso/capstone/domain/channel/dto/response/FavoriteVideoInfoResponse.java
@@ -36,8 +36,8 @@ public record FavoriteVideoInfoResponse (
     ) {}
 
     public record SentimentDistribution(
-            Double positive,
-            Double negative,
-            Double other
+            Integer positive,
+            Integer negative,
+            Integer other
     ) {}
 }
